add user module

+++ maimaiti/collection/views.py
@@ -8,22 +8,29 @@ from post.views import parse_html
 from django.http import HttpResponse
 import json
 
-def collect(request,user_id,post_id):
-    time = datetime.time
-    try:
-        Collection.objects.get(user_id = user_id,post_id = post_id)
+def collect(request,post_id):
+    if request.user.is_active:
+        user = request.user
+        time = datetime.time
+        try:
+            Collection.objects.get(user_id = user.id,post_id = post_id)
+            dic ={
+                'flag':0,
+                'msg':'您已收藏!',
+            }
+        except Collection.DoesNotExist:
+            collection = Collection(user_id = user.id,post_id = post_id,time = time)
+            collection.save()
+            collections = Collection.objects.filter(post_id = post_id)
+            dic ={
+                'flag':1,
+                'msg':'收藏成功!',
+                'follower':len(collections),
+            }
+    else:
         dic ={
-            'flag':0,
-            'msg':'您已收藏!',
-        }
-    except Collection.DoesNotExist:
-        collection = Collection(user_id = user_id,post_id = post_id,time = time)
-        collection.save()
-        collections = Collection.objects.filter(post_id = post_id)
-        dic ={
-            'flag':1,
-            'msg':'收藏成功!',
-            'follower':len(collections),
+            'flag':-1,
+            'msg':'请先登录!',
         }
     dic_json = json.dumps(dic)
     response=HttpResponse()
@@ -31,20 +38,27 @@ def collect(request,user_id,post_id):
     response.write(dic_json)
     return HttpResponse(response)
 
-def cancel_collect(request,user_id,post_id):
-    try:
-        collection = Collection.objects.get(user_id = user_id,post_id=post_id)
-        collection.delete()
-        collections = Collection.objects.filter(post_id = post_id)
+def cancel_collect(request,post_id):
+    if request.user.is_active:
+        user = request.user
+        try:
+            collection = Collection.objects.get(user_id = user.id,post_id=post_id)
+            collection.delete()
+            collections = Collection.objects.filter(post_id = post_id)
+            dic ={
+                'flag':1,
+                'msg':'取消收藏成功!',
+                'collection':len(collections),
+            }
+        except Collection.DoesNotExist:
+            dic ={
+                'flag':0,
+                'msg':'未收藏此文章!',
+            }
+    else:
         dic ={
-            'flag':1,
-            'msg':'取消收藏成功!',
-            'collection':len(collections),
-        }
-    except Collection.DoesNotExist:
-        dic ={
-            'flag':0,
-            'msg':'未收藏此文章!',
+            'flag':-1,
+            'msg':'请先登录!',
         }
     dic_json = json.dumps(dic)
     response=HttpResponse()
@@ -64,5 +78,5 @@ def listing(request,user_id):
                 post_list.append(post)
             except Post.DoesNotExist:
                 pass
-    return render_to_response('list4.html',{'post_list_0':post_list,'flag':'1'})
+    return render_to_response('list4.html',{'post_list_0':post_list,'flag':'1','user':request.user})
